refactor: create action folder
