+++ src/components/profileSetting/DeleteAccountButton.tsx
@@ -1,6 +1,5 @@
 'use client';
 import React from 'react';
-import { Button } from '../ui/button';
 import {
   Dialog,
   DialogClose,
@@ -11,9 +10,10 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { DeleteAccount } from './action';
 import { redirect } from 'next/navigation';
 import { signOut } from 'next-auth/react';
+import { Button } from '@/components/ui/button';
+import { DeleteAccount } from '@/action/setting/action';
 
 const DeleteAccountButton = ({ access_token }: { access_token: string }) => {
   async function onSubmit() {
